refactor(api): move home to new pattern
